chatbot/requirements.txt
fastapi
uvicorn
pydantic
websockets
python-multipart
requests
json-repair
python-dotenv
crewai
ollama